add more integration tests, move datavalues
