+++ lib/src/index.ts
@@ -4,3 +4,4 @@ export * from './tracker';
 export * from './searcher';
 export * from './factory';
 export * from './models/data-contracts';
+export * from './models/dataValue';
